+++ frontend/src/scenes/organization/membersLogic.tsx
@@ -4,15 +4,13 @@ import { loaders } from 'kea-loaders'
 
 import api from 'lib/api'
 import { OrganizationMembershipLevel } from 'lib/constants'
-import { dayjs } from 'lib/dayjs'
 import { lemonToast } from 'lib/lemon-ui/LemonToast/LemonToast'
-import { capitalizeFirstLetter, fullName } from 'lib/utils'
 import { permanentlyMount } from 'lib/utils/kea-logic-builders'
 import { membershipLevelToName } from 'lib/utils/permissioning'
 import { organizationLogic } from 'scenes/organizationLogic'
 import { userLogic } from 'scenes/userLogic'
 
-import { ExporterFormat, OrganizationMemberScopedApiKeysResponse, OrganizationMemberType } from '~/types'
+import { OrganizationMemberScopedApiKeysResponse, OrganizationMemberType } from '~/types'
 
 import type { membersLogicType } from './membersLogicType'
 
@@ -20,62 +18,12 @@ export interface MembersFuse extends Fuse<OrganizationMemberType> {}
 
 const PAGINATION_LIMIT = 200
 
-function escapeCsvField(value: string): string {
-    if (/[",\n\r]/.test(value)) {
-        return `"${value.replace(/"/g, '""')}"`
-    }
-    return value
-}
-
-function yesNo(condition: boolean): string {
-    return condition ? 'Yes' : 'No'
-}
-
-function buildOrganizationMembersCsv(members: OrganizationMemberType[]): string {
-    const headers = [
-        'Name',
-        'Email',
-        'Organization role',
-        '2FA enabled',
-        'Email verified',
-        'Social login',
-        'Joined at',
-        'Last login',
-        'User UUID',
-    ]
-
-    const lines = [
-        headers.map(escapeCsvField).join(','),
-        ...members.map((member) => {
-            const roleLabel = capitalizeFirstLetter(
-                membershipLevelToName.get(member.level) ?? `unknown (${member.level})`
-            )
-            const emailVerified = member.has_social_auth || !!member.user.is_email_verified
-            const row = [
-                fullName(member.user),
-                member.user.email,
-                roleLabel,
-                yesNo(member.is_2fa_enabled),
-                yesNo(emailVerified),
-                yesNo(member.has_social_auth),
-                member.joined_at,
-                member.last_login ?? '',
-                member.user.uuid,
-            ]
-            return row.map((cell) => escapeCsvField(String(cell))).join(',')
-        }),
-    ]
-
-    return lines.join('\r\n')
-}
-
 export const membersLogic = kea<membersLogicType>([
     path(['scenes', 'organization', 'membersLogic']),
     connect(() => ({
-        values: [userLogic, ['user'], organizationLogic, ['currentOrganization']],
+        values: [userLogic, ['user']],
     })),
     actions({
-        downloadMembersList: true,
         ensureAllMembersLoaded: true,
         loadAllMembers: true,
         loadMemberUpdates: true,
@@ -213,15 +161,6 @@ export const membersLogic = kea<membersLogicType>([
                 return Math.max(count ?? 0, members?.length ?? 0)
             },
         ],
-        downloadMembersListDisabledReason: [
-            (s) => [s.membersLoading, s.sortedMembers],
-            (membersLoading, sortedMembers): string | null =>
-                membersLoading && !sortedMembers?.length
-                    ? 'Loading members…'
-                    : !sortedMembers?.length
-                      ? 'No members to export'
-                      : null,
-        ],
     }),
 
     listeners(({ values, actions }) => ({
@@ -229,31 +168,6 @@ export const membersLogic = kea<membersLogicType>([
             if (userUuid === userLogic.values.user?.uuid) {
                 location.reload()
             }
-        },
-
-        downloadMembersList: async () => {
-            const sortedMembers = values.sortedMembers
-            if (!sortedMembers?.length) {
-                lemonToast.warning(values.downloadMembersListDisabledReason ?? 'No members to export')
-                return
-            }
-            // Dynamic import avoids a static cycle: membersLogic → exportsLogic → … → hedgehogModeLogic → membersLogic
-            const { exportsLogic } = await import('lib/components/ExportButton/exportsLogic')
-            const csv = buildOrganizationMembersCsv(sortedMembers)
-            const slug = (values.currentOrganization?.name ?? 'organization')
-                .replace(/[^a-z0-9]+/gi, '_')
-                .replace(/^_|_$/g, '')
-                .toLowerCase()
-                .slice(0, 64)
-            const filename = `posthog_organization_members_${slug || 'organization'}_${dayjs().format('YYYY-MM-DD')}.csv`
-            exportsLogic.actions.startExport({
-                export_format: ExporterFormat.CSV,
-                export_context: {
-                    localData: csv,
-                    filename,
-                    mediaType: ExporterFormat.CSV,
-                },
-            })
         },
 
         ensureAllMembersLoaded: async () => {
