+++ frontend/src/scenes/settings/organization/Members.tsx
@@ -23,6 +23,7 @@ import {
     organizationMembershipLevelIntegers,
 } from 'lib/utils/permissioning'
 import { twoFactorLogic } from 'scenes/authentication/twoFactorLogic'
+import { membersExportLogic } from 'scenes/organization/membersExportLogic'
 import { membersLogic } from 'scenes/organization/membersLogic'
 import { organizationLogic } from 'scenes/organizationLogic'
 import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
@@ -184,15 +185,20 @@ function ActionsComponent(_: any, member: OrganizationMemberType): JSX.Element |
 }
 
 export function Members(): JSX.Element | null {
-    const { downloadMembersListDisabledReason, filteredMembers, membersLoading, search } = useValues(membersLogic)
+    const { filteredMembers, membersLoading, search } = useValues(membersLogic)
+    const { downloadMembersListDisabledReason } = useValues(membersExportLogic)
     const { currentOrganization } = useValues(organizationLogic)
     const { preflight } = useValues(preflightLogic)
     const { user } = useValues(userLogic)
-    const { downloadMembersList, setSearch, ensureAllMembersLoaded } = useActions(membersLogic)
+    const { setSearch, ensureAllMembersLoaded } = useActions(membersLogic)
+    const { downloadMembersList } = useActions(membersExportLogic)
     const { updateOrganization } = useActions(organizationLogic)
     const { openTwoFactorSetupModal } = useActions(twoFactorLogic)
 
     const twoFactorRestrictionReason = useRestrictedArea({ minimumAccessLevel: OrganizationMembershipLevel.Admin })
+    const downloadMembersListRestrictionReason = useRestrictedArea({
+        minimumAccessLevel: OrganizationMembershipLevel.Admin,
+    })
     const membersCanInviteRestrictionReason = useRestrictedArea({
         minimumAccessLevel: OrganizationMembershipLevel.Admin,
     })
@@ -333,14 +339,16 @@ export function Members(): JSX.Element | null {
                     onChange={setSearch}
                     className="flex-1 basis-[min(100%,18rem)]"
                 />
-                <LemonButton
-                    type="secondary"
-                    onClick={downloadMembersList}
-                    disabledReason={downloadMembersListDisabledReason}
-                    data-attr="org-members-download-csv"
-                >
-                    Download members list
-                </LemonButton>
+                {!downloadMembersListRestrictionReason && (
+                    <LemonButton
+                        type="secondary"
+                        onClick={downloadMembersList}
+                        disabledReason={downloadMembersListDisabledReason}
+                        data-attr="org-members-download-csv"
+                    >
+                        Download members list
+                    </LemonButton>
+                )}
             </div>
 
             <LemonTable
